Move method to seperate java class
