+++ deduplication/src/main/java/gov/cdc/nbs/deduplication/duplicates/service/MergeGroupHandler.java
@@ -20,13 +20,16 @@ public class MergeGroupHandler {
 
   private final NamedParameterJdbcTemplate deduplicationTemplate;
   private final PatientRecordService patientRecordService;
+  private final MergeGroupService mergeGroupService;
 
   public MergeGroupHandler(
       @Qualifier("deduplicationNamedTemplate") NamedParameterJdbcTemplate deduplicationTemplate,
-      PatientRecordService patientRecordService
+      PatientRecordService patientRecordService,
+      MergeGroupService mergeGroupService
   ) {
     this.deduplicationTemplate = deduplicationTemplate;
     this.patientRecordService = patientRecordService;
+    this.mergeGroupService = mergeGroupService;
   }
 
   public List<MatchesRequireReviewResponse> getPotentialMatches(int page, int size) {
@@ -126,7 +129,7 @@ public class MergeGroupHandler {
   }
 
   public List<MatchesRequireReviewResponse> getAllMatchesRequiringReview() {
-    List<MatchCandidateData> candidates = patientRecordService.fetchAllMatchesRequiringReview();
+    List<MatchCandidateData> candidates = mergeGroupService.fetchAllMatchesRequiringReview();
 
     return candidates.stream()
             .map(candidate -> {
